+++ docs/releases.md
@@ -14,6 +14,7 @@
 * Add utility function `get_glorys_bounds` to return the min/max latitude and longitude required for GLORYS downloads for a fixed domain ([#445](https://github.com/CWorthy-ocean/roms-tools/pull/445))
 * Add `compute_total_cdr_source ` method to `CDRForcing`, which computes integrated tracer quantities for all releases and returns a pandas DataFrame including a units row ([#444](https://github.com/CWorthy-ocean/roms-tools/pull/444))
 * Plot ensembles of CDR uptake curves and their mean and standard deviation ([#467](https://github.com/CWorthy-ocean/roms-tools/pull/467))
+* Allow custom shapefiles (such as GSHHS) to be used for coastline masking ([#475](https://github.com/CWorthy-ocean/roms-tools/pull/475))
 
 ### Breaking Changes